+++ src/modules/tools/routing/store/actionsRouting.js
@@ -7,6 +7,7 @@ import crs from "@masterportal/masterportalapi/src/crs";
 import {fetchRoutingBkgGeosearch, fetchRoutingBkgGeosearchReverse} from "../utils/geosearch/routing-bkg-geosearch";
 import {fetchRoutingLocationFinderGeosearch} from "../utils/geosearch/routing-locationFinder-geosearch";
 import {fetchRoutingKomootGeosearch, fetchRoutingKomootGeosearchReverse} from "../utils/geosearch/routing-komoot-geosearch";
+import {fetchRoutingElasticGeosearch} from "../utils/geosearch/routing-elastic-geosearch";
 import * as constantsRouting from "./constantsRouting";
 
 /**
@@ -76,6 +77,9 @@ export default {
             else if (state.geosearch.type === "KOMOOT") {
                 geosearchResults = await fetchRoutingKomootGeosearch(search);
             }
+            else if (state.geosearch.type === "ELASTIC") {
+                geosearchResults = await fetchRoutingElasticGeosearch(search);
+            }
             else {
                 throw new Error("Geosearch is not configured correctly.");
             }
@@ -83,7 +87,7 @@ export default {
             // Transform Coordinates to Local Projection
             geosearchResults.forEach(async geosearchResult => {
                 if (!geosearchResult.getEpsg()) {
-                    geosearchResult.setEpsg("4326");
+                    geosearchResult.setEpsg(state.geosearch.epsg);
                 }
                 const coordinatesLocal = await dispatch(
                     "Tools/Routing/transformCoordinatesToLocalProjection",
@@ -126,6 +130,9 @@ export default {
             }
             else if (state.geosearch.type === "KOMOOT") {
                 geosearchResults = await fetchRoutingKomootGeosearchReverse(coordinates);
+            }
+            else if (state.geosearch.type === "ELASTIC") {
+                geosearchResults = null;
             }
             else {
                 throw new Error("Geosearch is not configured correctly.");