+++ src/modules/tools/routing/store/stateRouting.js
@@ -35,7 +35,10 @@ const state = {
         // type: "BKG",
         // serviceId: "bkg_geosearch"
         type: null,
-        serviceId: null
+        serviceId: null,
+        epsg: "4326",
+        searchField: null,
+        sortField: null
     },
     geosearchReverse: {
         // type: "NOMINATIM",